+++ CHANGES.md
@@ -9,7 +9,7 @@
 ## 0.5.0 / 2017-January-??
 
 * [TODO] Instrumentation logging event for SQL-execution should include `ISqlSource` instead of SQL string
-* [TODO] Overhaul `asphalt.core/defsql`
+* Overhaul `asphalt.core/defsql`
   * Named by default
     * Implement `clojure.lang.Named` interface, hence `(name sql-source)` works
     * Name can be overridden via `:sql-name` option kwarg
@@ -20,8 +20,8 @@
     * Param setter  : auto-default to use param types when specified
     * Row maker     : auto-default to use result types when specified
     * Column reader : default is used
-    * [TODO] Connection worker : default to `query` vs `update` based on first SQL token
-  * [TODO] Behave as function `(f connection-source params)` - with associated connection-worker
+    * Connection worker : default to `query` vs `update` based on first SQL token
+  * Behave as function `(f connection-source params)` using associated connection-worker
 * Transactions
   * Do not override the exception causing rollback/commit by the exception in rollback/commit
 * Types/Protocols
@@ -45,6 +45,7 @@
   * [BREAKING CHANGE] Accept options in `asphalt.core` fetch fns (arity 3)
   * [BREAKING CHANGE] Drop `asphalt.core/defquery` in favor of `asphalt.core/defsql` behaving as function
   * Helper fn `asphalt.core/default-fetch` for default values in single row access
+  * [TODO] Re-implement `asphalt.core/fetch-maps` with options
 
 
 ## 0.4.0 / 2015-November-30